requirements.txt
streamlit
python-dotenv
pandas
plotly
finance-datareader
requests
beautifulsoup4
websocket-client
playwright
